Address PR #513 review: stop leaking temp paths, dedupe fixture path helpers
- Drop the raw fs error text from the two temp-copy fixture-resolution
  errors in runner.ts. It embedded absolute os.tmpdir() paths, leaking the
  environment and making the offline eval's diagnostics non-deterministic.
  The normalized repo-relative path in the message already names the fixture.
- Extract the duplicated local-fixture path validators (normalize / under-area
  / contained-in) and the fixture-areas map into evals/lib/fixture-paths.ts,
  shared by runner.ts and scenario-loader.ts. This removes the duplication and
  the drift between the two copies (the runner's rejected Windows drive-letter
  paths the loader's accepted); both now use the stricter version.
